Update  presque fini

+++ src/pokemons-cards/pokemons-cards.controller.ts
@@ -1,11 +1,12 @@
 import{ Request, Response } from 'express';
 import prisma from '../client';
-
+import { error } from 'console';
+import jwt from 'jsonwebtoken';
  
 
 export const getPokemons = async (_req: Request, res: Response) => {
     const pokemons = await prisma.pokemonCard.findMany();
-    res.status(200).send(pokemons);
+    res.status(200).send({ success : pokemons}); 
 }
 
 export const getPokemonById = async (req: Request, res: Response) => {
@@ -14,18 +15,18 @@ export const getPokemonById = async (req: Request, res: Response) => {
         where: { id: parseInt(pokemonCardId) },
     });
     if (pokemon) {
-        res.status(200).send(pokemon);
+        res.status(200).send({ success : pokemon});
     } else {
-        res.status(404).send(`Pokémon avec l'ID ${pokemonCardId} non trouvé`);
+        res.status(404).send({error : `Pokemon avec ID ${pokemonCardId} non trouve`});
     }
 };
 
 export const createPokemon = async (req: Request, res: Response): Promise<void> => {
     const { name, pokedexId, typeName, lifePoints, size, weight, imageUrl } = req.body;
 
-    // Vérifiez si typeName est défini
+    // Verifiez si typeName est defini
     if (!typeName) {
-        res.status(400).send('Le champ typeName est requis');
+        res.status(400).send({ error : 'Le champ typeName est requis'});
         return;
     }
 
@@ -35,7 +36,7 @@ export const createPokemon = async (req: Request, res: Response): Promise<void> 
     });
 
     if (!type) {
-        res.status(404).send(`Type ${typeName} non trouvé`);
+        res.status(404).send({error : `Type ${typeName} non trouve`});
         return;
     }
 
@@ -61,18 +62,44 @@ export const createPokemon = async (req: Request, res: Response): Promise<void> 
 export const editPokemon = async (req: Request, res: Response) => {
     const { pokemonCardId } = req.params;
     const updatedProperties = req.body;
+
+    // Verifie si la carte Pokemon existe
+    const existingPokemon = await prisma.pokemonCard.findUnique({
+        where: { id: parseInt(pokemonCardId) },
+    });
+
+    if (!existingPokemon) {
+        res.status(404).send({ error: `Pokemon avec ID ${pokemonCardId} non trouve` });
+        return;
+    }
+
+    // Si la carte existe, met à jour
     const updatedPokemon = await prisma.pokemonCard.update({
         where: { id: parseInt(pokemonCardId) },
         data: updatedProperties,
     });
-    res.status(200).send(updatedPokemon);
+
+    res.status(200).send({ success: updatedPokemon });
 };
 
 
 export const deletePokemon = async (req: Request, res: Response) => {
     const { pokemonCardId } = req.params;
+
+    // Verifie si la carte Pokemon existe
+    const existingPokemon = await prisma.pokemonCard.findUnique({
+        where: { id: parseInt(pokemonCardId) },
+    });
+
+    if (!existingPokemon) {
+        res.status(404).send({ error: `Pokemon avec ID ${pokemonCardId} non trouve` });
+        return;
+    }
+
+    // Supprime la carte Pokemon
     await prisma.pokemonCard.delete({
         where: { id: parseInt(pokemonCardId) },
     });
-    res.status(200).send(`Pokémon avec l'ID ${pokemonCardId} supprimé`);
+
+    res.status(200).send({ success: `Pokemon avec ID ${pokemonCardId} supprime` });
 };
